perftests: continued infrastructure work
Use the Application class directly, instead of going through main.
Otherwise main will parse the commandline a second time, and it is
also a nice simplification. And use decorators for the GUI tests.

Removed test teardown code, as it is very error-prone trying to make
the tests not influence each other.