Use event timestamps to measure key press duration
When pressing home shortly to toggle layer solo mode, there was/is
a problem when the display update takes too much time. We then
sometimes get a too long key press reported, and we leave layer solo
again because we think the user was holding the key down intentionally
long.