brushgroups implemented, basic drag&drop working (wip, known bugs)
Partially based on the implementation by Ilya Portnov. One major
difference is that groups are not stored into the brush itself.
This avoids duplication of information, and needless copies of
stock brushes (which are not updated any more on changes).

Removed the application's brush list in favour of the group dict.
Logic for merging changes in the shipped brushes was rewritten.

TODO:
- GUI to add groups
- brush management (file creation/deletion) known to be buggy
- fix unittest