brushgroups: fix unicode problem
Somehow when starting mypaint from emacs my system's default
encoding is set to "ascii". This made the current implementation
throw exceptions when using non-ascii group names.

Seems safer to write our config file in a fixed encoding (utf-8)
instead of using the system's default (which might change, or might
not be able to encode one of the names of the default group names).