colorsampler: remove entry focus hack
The code was not functional code because it was only partially merged in.

This was a workaround for the problem that pressing numbers inside the
entry does not work. However I think the usecase is not important enough
for such a complicated workaround (and there most likely is a better one).