brushes: brushset update from Ramón
"I have made some new brushes to add to my set. The Reason is that we can
now have a good dynamic glow airbrush,( untill blending modes appear) So i
have mapped saturation, lightness and hue changes to increase the fake
effect.  more pressure hue changes and it looks like additive mode, it works on
my tests so there you have. B-pencil And H-Pencil are made for people who
like draw skethes but not so dirty. they produce a clean stroke that we can
control more than a dirty brushstroke. good for Architecture or comic
drawings , like superman on the forum. Then we have a cool Thin Pen that
is v5 pilot pen . with a lot of things that can be easily manipulated to
vary the effect.. What about blenders? yes! i like a lot blenders and this is
the reason that i have made some more  again. Water, is usefull to blend
colors like we do with a simple brush with water and no color. and the others
are more expressive for weird and very funny effects. They Work Well at high
res. and i hope you will enjoy them.maybe the last brushes for a time,
(allwais saying the same thing  ¬¬)"