strokemap: optimize save_ora, get rid of protobuf library
Most of the saving time was spent inside protobuf while saving the
strokemap.  We now use a different binary format. Old strokemaps
in existing files will be discarded.

In addition, about 50% of the strokemap size and saving time
was used for the brush settings. Those were saved as string with
each stroke. We now save only an index in a separate brush table.