Windows: workaround for bad timer resolution
Fix for 0.9 release; fixing it in the brush engine would require
careful checking/retuning of existing brushes.
https://gna.org/bugs/index.php?16569

This also fixes far-out dab with some brushes on Windows that would
enlarge the canvas size a lot, possibly leading to out-of-memory
when saving. This was happening because the brush engine uses a very
small time delta (0.0001 seconds) when the dtime input is zero.