thumbnails: 256x256 preview, 20% faster saving
Some refactoring of the thumbnailing code. We now use 256x256
whenever possible. When saving a PNG, we no longer load the full
file back from disk (in addition to the data we already hold in
memory) just to generate the thumbnail.

Saving ORA and PNG is both around 20% faster now. This is not
noticed by the performance benchmarks that don't use the GUI.