windows: code cleanup
Use raw string for backslashes.
Fix exception handling. As it was before, in the case
of an exception the use of "localappdatapath" later would
have triggered another (fatal) exception.
Don't bother closing handles manually, python destructors do this
for us.
Set the path on windows independent of the result of expanduser('~').