remove "about brush" button
The problem of attributing brushes still needs to be solved, but
this solution didn't work at all. Nobody wants to click a "info"
button just to find out it there is no info available, or to see
the same info again for every brush in the same category.

The external editor didn't work well for me on Linux, neither.

On Windows it is completely broken: https://gna.org/bugs/?17353