layers: fix rename dialog not showing sometimes
Detect clicks on layers in the dialog using their glib type, and respond
appropriately on the corresponding button-up (for consistency). The user
must still click once to select a layer before it'll become selected,
and thus eligable for renaming with a double click, so our text is in
fact incorrect right now.

https://gna.org/bugs/?17656