layout: positioning fixes
No longer use gtk.WIN_POS_MOUSE: https://gna.org/bugs/?17915

Favour gtk.Window.parse_geometry if we have all the required window
position and size info. This works better with some window managers.

Definitively set the position of tools which are snapped out by
dragging with the mouse and the preview window to take of the preview
window in a deferred handler.