Don't translate a sentence fragment.
It's not a good idea to translate sentence fragments, especially ones
involving plurals. Depending on the language, the whole sentence might
translate differently (e.g., French plural verbs). This also translates
formatting, which is probably not a good idea.

I tried to combine the translations as best I could, but some of them
weren't translated completely. Some were fuzzy, also, which is probably
because of the sentence fragment, but I'm just guessing.