refactor brush observer code and its users
Make ManagedBrush the primary class for storing, manipulating
and observing brush settings.

The Brush class (the one which inherits from the C++ brush
class and is used for painting on a surface) is no longer
manipulated directly, but observes a ManagedBrush instead.

Refactor functionwindow.py such that it no longer assumes
to be the only manipulator of app.brush. It will now update
its GUI when the brush is modified, not only when a different
brush gets selected.

ManagedBrush now owns a private dictionary instead of
inheriting from dict, and is manipulated from outside
through accessor functions (to make it observable).
Observing the dict alone is not enough, because it contains
list instances that are manipulated without modifying the dict.

startup performance benchmark: no change