undo: slightly increase number of steps
Moderate increase from 20 to 30 steps.

Seems some people thought we have too few steps now, which is
probably because we started to split strokes more often.

Memory usage of a step can be significant if the stroke goes
all over the image. Ideally we would take the diff and compress
the history during natural breaks, like with the strokemap...