layout: floating tools are now undecorated
Remove the window manager decorations from tool windows in the floating
state. This avoids witlessly repeating the title and close button at the
top of the tool, and it saves a bit of vertical space too. Also the
snap-out drag preview rectangle now shows you exactly where the window
will go when you drop the window. To me it all looks and feels a little
tighter and more consistent.

Possibly controversial because people choose window managers for the
looks and behaviour they like, but this commit should be easy to revert
since it's just one line. OTOH, not every WM will respect the
decoratedness hint anyway, so we may be fine there. The windows are
still managed normally, so alt-tab should work normally where it's
already supported.