brushmanager: fix brush saving after device change
Make selected_brush always the same as the highlighted brush.

When selecting a devbrush (because of switching devices), we used
to highlight the parent brush, but behind the scenes the temporary
devbrush was selected. In the brush editor, saving the brush settings
would save into the temporary devbrush, instead of the selected brush.
Reported here:
http://forum.intilinux.com/mypaint-general-discussion/brushes-not-saving/

Remove find_brushlist_ancestor(). We don't need to do a deep search,
because the parent of a brush, if it exists, is always a ManagedBrush.