layers tool: redo as a gtk list
Re-express the layers list as a GTK ListStore and TreeView combination
rather than as a custom widget. Layers can now be dragged around within
the list to reorder them, and the list scrolls to keep up when the user
moves the current layer out of sight using the keys or buttons.

https://gna.org/bugs/?18302
https://gna.org/bugs/?18304