brushmodifier: redo states using gtk.ToggleActions
It's fiddly to have toolbar and menu toggleactions track the state of an
internal StateGroup's states, so rewrite brushmodifier using normal
gtk.Actions.

We now distingush between brushes that are dedicated eraser brushes and
ones which aren't when entering and leaving eraser mode.

This emulates a group of radio items, but since two of them can be
toggled off by clicking them or pressing their key, it makes sense for
them to be ToggleActions instead.