Scratchpad: Refactored palette drawing code for reuse
Moved the code to lib/scratchpad_palette.py:

squiggle, *_squiggle --> various methods to generate brush 'events' that
draw different 'squiggles'.

draw_palette --> parameterised method to draw the palette. Can be given any document.Document
object to draw on, and can be passed a palette of colours, offset and squiggle method to use.

GimpPalette --> palette that can load in a gimp palette. May need to be renamed to simply 'Palette' and
change 'load' method to reference gimp format as this palette object is more generic