app: use an explicit path for icon searches
Create the app instance with an explicit extra location for the icons,
based in the install prefix as derived from argv[0]. Fixes the app not
beeing able to start sensibly if installed somewhere not within the
bases for the usual icon search path. Be even more picky about missing
needed icons too so that people doing odd things get nicer feedback.

https://gna.org/bugs/?18442