toolbar: ignore really huge style defaults
Override user settings of gtk.TOOLBAR_BOTH for the toolbar becuase
that's way too big for sensible display of the number of buttons we
have even at this early stage. Use gtk.TOOLBAR_BOTH_HORIZ instead since
that partly respects the user's desire for text (or it would if we
declared any items as important).

gtk.TOOLBAR_BOTH_HORIZ is the default for most systems these days.