toolbar: force into icons-only mode
Workaround for a (py)GTK bug. The style-changed signal fires at startup
with an erroneous GTK_TOOLBAR_BOTH setting even when overridden both in
the user's gtkrc and in gconf, and even when the docs say the default is
GTK_TOOLBAR_BOTH_HORIZ. Forcing a style at that point means we can't
respect the user's style changes later, and rather than fiddle and set
up handlers within handlers and state flags to model GTK's peculiar
world view, it's simpler for our code to just do something that'll be
right for most users (who yes, tend to be graphical thinkers with good
eyesight; but for those who aren't: we still allow overriding the
toolbar icon size, and tooltips still work).