pixops: dithering for display and for saving PNG
Since dithering is a cheap operation now, also enable it for
the display when converting from 15bpc to 8bpc. When saving
a solid PNG it is now dithered, too (not only transparent PNGs).

I'm not claiming I can see the difference on my display, but
you can certainly see when you post-process the PNG.

scroll_nozoom_onelayer: performance seems to be unchanged (<5%)