colorsel: support older pygtk, add x2click details
Since the PPA builds (and Windows builds if you want pressure support)
need to support an ancient pygtk without a wrapped gtk.HSV, go back to
the nasty old trick of rummaging around in a donor gtk.ColorSelection
for spare parts. Sigh. Featurewise, this is slightly better. We regain
draggable colours from the prev and next widgets, which we can use
elsewhere: bits of the UI are now sensitive to having colours dropped on
them, so might as well provide a source.

Also add a trivial but nice to have colour details modal dialog on a
double-click of the colour comparison boxes. You can now type hex codes
into the dialog to set the current colour.