change warmer/colder keyboard shortcut
Changed the warmer/colder keyboard shortcut to not stop at a designated
"cold/warm point" but instead go all the way around the hue circle.

The reason for this is simple -- combined with brighter/darker and
purer/grayer these keyboard shortcuts make for a very powerful way to
select colour using the keyboard without having to click in the selector
-- something I've not seen in any other program. With the hue selection
(cooler/warmer is not really a good name for it in this case -- maybe
 hue plus/minus?) wrapping around, I can put the colour selector on a
different screen entirely and only use it as a visual cue when
navigating it via the keyboard.