Menubar hiding, menu button on the toolbar
To save vertical space, the menu bar can now be hidden in the same way
as the toolbar. When the menu bar is hidden, a menu button appears on
the toolbar. This is roughly the way newer Firefox versions do it.

The defaults are now to show the toolbar and its menu button, and hide
the menubar.