app, layout: initial pos updates, better defaults
Give the Scratchpad a better initial position, and update the other
positions to make better use of screen space. All tools are initially
hidden since we now have a way of changing both brush and colour very
immediately on the toolbar. Hopefully that's sensible initial behaviour.

All windows now have an explicit "hidden" in prefs, and default values
are now chosen from the prefs in the same way everywhere it matters for
first display.

https://gna.org/bugs/?18518
https://gna.org/bugs/?18519