brush manager: abstract out getting parent brush
Rather than requiring other modules to fumble around with parent brush
names, export a method for getting a fully initialized parent based on a
BrushInfo or ManagedBrush object.

We should really do something about the number of repetitions of "load
this MangedBrush if it's on disk and hasn't been loaded yet" that are
becoming increasingly scattered around the code. That may involve
reworking/incorporating the existing reload-if-changed method. Mark as
dubious/XXX for now.