toolbar: quick brush chooser on dropdown & default
The quick brush chooser is now exposed on the panel directly as part of
a new dropdown panel. The brush dropdown button is no longer in two
parts: testing showed that people didn't really understand the reason
for having it in two bits, so let's simplify it.

Convert the quick brush chooser's brush list combobox to a custom
spinner-like widget; done for pointer grab related reasons in the
dropdown. Hopefully it's a fairly understandable control and easy to hit
with a stylus.

Expose the brush chooser as an Action, and make it the primary method of
choosing a brush; the default binding for "b". This is intended to bring
it closer in feel to a popup chooser like those for colour, which makes
sense to me since its functionality is beginning to approach that of a
popup menu (and we should probably go further).

The existing Brush List has moved to "<shift>b" by default and is
renamed to "Brush List Editor": you can still use it as as a permanent
fixture for choosing brushes though. Rename the settings editor for
consistency too, but don't change its default "<control>b" binding.

Testers were complaining about the height of the toolbar, so drop it back
to the usual GTK_ICON_SIZE_LARGE_TOOLBAR size. The new graphical
tooltips should help make the brush preview useful.