glade: tileddrawwidget support for glade projects
The TiledDrawWidget is a fairly simple canvas, so nothing much needs to
be changed to support using it within glade projects. It has a slightly
dubious need to refer to a zoom quality setting, and a standalone model
needs to be wired in to support drawing within glade (very minimally :)
but that's about it.