toolbar: menu for selecting which tools are wanted
Add a right-mouse-button menu on the toolbar for choosing which sets of
tools are required. Settings are saved in the preferences.

Reinstate blend modes on the toolbar (and the "cute pink eraser" icon
:-) for those who want them. They're off by default.

Add buttons for scrapbook saving and Prev/Next Scrap to the toolbar,
switched off by default. Interim icons for Save/Prev/Next Scrap. They're
from stock, and could probably be improved.

Trying fullscreen/unfullscreen on the top-right. It feels like a nice
place to put things that make major but reversible changes to the gui,
and it's a nice Fitts's-compliant use of the space.

No togglability for the dropdowns yet, given the fiddly setup
post-construction.