Make TiledSurface an internal member of Layer
Most part of the code should not care about
the surface implementation used. Provide wrappers
on layer for common things, and let the exotic
things call layer._surface to make it clear they
are dealing with internals.
In order to make it easier to swap out TiledSurface
with another surface implementation.