move layer: improved feedback & responsiveness
Improve the apparent responsiveness during a drag to move the current
layer. The idea is to update the display during the drag in an idle
handler, lazily to avoid doing unnecessary work. Schedule the first tile
shifts to happen around the cursor, since that's where the user is
presumably looking. Still a little slow to start because of the initial
snapshotting and sorting, but it feels far more responsive now.

Add cursors for the various drag operations, just because.

Hourglass/watch cursor after a move-layer drag while it slots the
remaining tiles into place.