scratchpad: update icons, fix icon sizes
The icon for the scratchpad was a bit too close to the new colorize icon
(which is a redraw of the Gimp's one, thus an older concept). Use an
older scratchpad icon I did a while back - lost the source icon sheet
unfortunately, but the bitmaps and scalable are still workable.

Coincidentally fixes https://gna.org/bugs/?19180