overlays: api for showing info on tdws
Experimental stab at an internal interface for animated or static
fragments of overlaid information which respond to user events.
Currently just a scale/zoom level indicator and a marker for the last
painting position, both of which fade out and are careful to only redraw
their own section of the screen.

Intended uses: textual explanations of fancy modes with lots of keyboard
chording, showing the document frame both normally and when being
edited, anything which must be displayed that isn't the canvas being
worked on. Maybe even previews of the current brush, its radius, colour,
stylus nib angle, whatever.