feedback/overlays: begin categorizing, add toggles
Add toggle actions and boolean user settings for those classes of
feedback which are currently implemented as overlays. Scale display
defaults to on, last-pos feedback defaults to off.

If we add a statusbar, certain kinds of textual feedback should go
there and not to the overlays if the statusbar is visible. Fading
overlays are quite visually noisy! This will require some sort of
textual feedback manager, but... later.

Regularize overlay constructors. Move responsibility for overlay
management to the drawwindow for now.