lcms2: transforms on load; groundwork for save
Basic build infrastructure for lcms2.

Load PNG files using whatever colour management info is in there, but
convert to sRGB 8-bit using lcms2. At the same time, detect PNGs which
were written without CM chunks.

If/when we go to linear light compositing in line with other OpenRaster
editors, we'll need to detect old OpenRaster files and do something sane
with them. One way of detecting would be to use lack of CM chunks as a
strong hint that a file was written by an old version of MyPaint. It'd
need confirmation; possibly using (the presence of) OpenRaster version
numbers in stack.xml assuming the spec dictates linear light at the same
time that MyPaint starts writing version numbers.

Infrastructure in place for saving PNG files with sRGB+cHRM+gAMA chunks,
but not utilized. Right now, MyPaint defaults to *not* saving them
because I might want to use this condition for a flag (see above...)