tdw: refactoring: encapsulate Cairo better
All of the external calls to get_model_coordinates_cairo_context() were
used for the purposes of converting between display and model
coordinates. If we decide to change the display backend at any point,
exposing this implementation detail might cause problems. Provide a pair
of methods for coordinate transforms instead, and begin using them. The
code becomes slightly more readable in places as a result.