TiledDrawWidget: Split into four different classes
Make the reponsibilities of this widget more clear by separating
out the two main aspects into separate classes: one for
rendering and one for event handling. A utility class is
used for drag handling and cursor updates.
In addition it makes the public API of TiledDrawWidget explicit.

This will hopefully make it easier to reuse the code that
is not backend-specific, like event and drag handling.