color sels: merge & revamp; palettes & gamut masks
Merge the old colour selectors into a single tabbed dockable with new
controls. Add an HCY wheel selector with a gamut mask. Add an editable
palette tab. Palettes and gamut masks can be saved as Gimp palette files
and loaded from them.

Major UI changes:

* The old dockable colour selectors have moved to a tabbed, dockable
  dialog: it's on the "c" key by defult, but this may change.

* The bottom of each tab has a picker button, a pair of previous/next
  comparison widgets, a button for adding the current brush colour to
  the palette, and a properties button for the current tab which is
  sometimes greyed out.

* The "HSV Cube" tab has a wide, squarish S+V surface like the centre of
  the old Color Sampler.

* Editable palettes. Load from, and save to GIMP palette files.

* The "HCY Wheel" tab displays hue, chroma, and perceptual luminance,
  and supports gamut masks.

Gamut masks are a new feature which replaces the old harmonies dial:
they're a significantly more flexible way of planning a colour scheme.
Presets for kinds of colour schemes which are similar to the old
harmonies are available via the New button in the properties dialog for
the HCY Wheel color adjuster.

The palette and gamut mask you were last working on are saved to the
MyPaint preferences regardless of whether you saved them to a file.

Internal changes:

* Submodule "colors", because these widgets can be run and tested
  independently from the MyPaint core. Should help with GTK3 porting, I
  guess.

* Manipulation of any colour shared between certain widgets now has its
  own interface.

* Colours manipulated via the UI now have specialized, interchangable
  and inter-convertible classes. It's normally better to use these then
  direct manipulation of floating-point HSV tuples.

* The brush colour is now best manipulated via app.brush_color_manager.
  Colour history now resides there.