brushlib: Add a trivial fixed-size MyPaintSurface implementation
The surface is backed by a single linear buffer, and has
a fixed size that is decided at creation time. It is intended
to be a simple example/reference implementation, not something
to use in production in non-trivial applications.
For those cases, MyPaintGeglTiledSurface can be used instead.

Note: currently untested.