palettes: default from file, add palettes dir
Add a palettes dubdiectory and begin populating it with palettes. Load
the initial palette from here when MyPaint starts up without a stored
palette prefs key.

The palettes subdir is installed as a datapath folder, i.e. to
/usr/share/mypaint/palettes. The files in there will typically be
read-only, but can be used as templates for new palettes. To make
finding them easier, make a shortcut for them in the palette load
dialog.

Minor docstring updates.