palette view: treat near-exact matches like exact
Extend the tolerance for "exact" matches a little because painted
strokes nominally of the same colour can vary in pixel colour to degrees
smaller than is easily perceptible, but which were still considered
approximate matches. The same is true of the edges of fuzzy strokes over
short distances.

Treating very near-exact matches like exact matches should improve the
flow of the PaletteNext and PalettePrev actions a little after a
pick-color action: if it's a near-exact match, you won't have to hit the
PaletteNext action twice to get the next swatch.