palette view: drag/drop and active area tweaks
Improve handling of cursor feedback.

The area with no swatches is now never treated as containing a colour.
This means that you have to actually click on swatches to select them,
but it makes dragging more understandable. Dragging to where there is no
swatch is now always a move (or copy) to the end of the palette.