Layer stack actions: update, make more keybindable
New keybindable commands in the menu:

  * Raising and lowering a layer in the layers stack
  * Duplicating the current layer
  * Renaming the current layer
  * Setting the current layer's visibility
  * Setting the current layer's locked status

Renaming a layer is now undoable, for consistency with the other layer
metadata commands.

Internal changes
----------------

A more sensible implementation for labelling unnamed layers. It's not a
fix for https://gna.org/bugs/?17991 yet, but that really isn't going to
be fixed until we decide not to allow unnamed layers at all, or save our
bodged-in reference numbers.

lib.command.RemoveLayer can now remove the final layer. If it does so,
it replaces it with a new unnamed layer. Required by the cleaner
nameless layer labelling code, which uses the layer to store its own
ref number, and it's better for identity and semantics. It was always
silly that "removing" the final layer would keep its name.