Add interaction modes for lines, view manipulation
Refactor LineMode to support the new modes interface.

Line modes are now accessible a) from the kitchen sink default freehand
mode with modifiers+drag, b) from the Edit menu and c) from the toolbar.
Other locations are possible too, of course.

I prefer all the sticky modes to be visible together on the toolbar
while they're the only reflection of what mode the app is in, but we're
admittedly getting a bit cluttered up there now. It's probably time for
some judicious culling and replacement.

New icons for rotating and zooming the view in "cursor and view" style,
and a generic Tango-style one for moving stuff.