fastpng: ignore non-RGB embedded ICC profiles
Gimp 2.8 happily embeds greyscale color profiles into PNG data with RGB
samples. This seems like a GIMP bug. Neverthless, it's revealed one of
ours: we need to be able to cope with non-RGB color profiles. Just
dropping them on the floor (similar to Gimp 2.6) should suffice while
we're using libpng to perform the limited number of transformations we
support.

https://gna.org/bugs/?20064