allow to choose if brushkeys remember color
It seems to be a tradition to change this behaviour every now and
then. Time for a preference (not in the preferences window, just
a toggle item in the menu). Also, update the current code to the
new brushmodifier behaviour.

Should we remember the color together with the brush?
Yes: http://gna.org/bugs/?15381
No:  http://gna.org/bugs/?16977
Yes: http://gna.org/bugs/?19548
Let's put an end to this insanity :-)