Make tile store interface in MyPaintTiledSurface concurrency friendly
Store a request state object in the consumer that wants to get/set tiles,
and allow tile stores to attach information to this object at the
start of a request, and use this to finalize the request on a subsequent
call.
The C++ interfaces used in MyPaint is not changed. Instead the wrapper
classes serializes requests so that one happens at a time.