layerswindow: reorder controls, better spacing
Now we have more blend modes, the mode GtkComboBox's menu can be cut off
short and awkwardly if the layers dialog is near the bottom of the
screen. Therefore move the controls which show details of the current
layer to the top, and the controls which affect the layers stack to the
bottom. The result is closer to the way the GIMP does it, and gives more
footroom for the combobox's menu. Headroom might now be a problem, but
layers aren't created using a mode that's further down the list: Normal
is at the top.

Space out controls in the Layers subwindow a tiny amount for some
breathing room. If this seems right, we should use it in other dockable
subwindows for consistency.