SetLayerOpacity: don't recalc bbox needlessly
Minor UI jerkiness removal for a special case. If SetLayerOpacity makes
no changes to the effective opacity - which should only happen for
invisible layers - the command now no longer recalculate the bbox of the
layer once or twice for each move of the opacity slider. This can add up
to an appreciable but small delay for layers with a lot of tiles, and
jerkiness in what could be a smooth slider.

https://gna.org/bugs/?20287

The problem really lies with tiledsurface.get_tiles_bbox() and
MyPaintSurface.get_bbox() iterating over index-tuple-keyed hashes merely
to obtain four numbers - would it help matters elsewhere to maintain a
cache of extents?