canvasevent: improve flow for spring-loaded modes
Make spring-loaded modes actually persist for as long as modifiers are
held. Not bouncing between modes will be better for UI feedback.

One little benefit: (any modifiers)+(toolbar button click) now engages a
"secret" versions of some modes which lasts only for as long as the
modifiers are held. For frame manipulation this means the dialogless
version of the mode. Really the "secret" version is just what you get
when you bind it to a (modified) button press via preferences.

Fix: releasing modifiers while dragging in oneshot+spring-loaded modes
should not exit the mode. That now happens correctly when the drag
finishes.