brushsettings: use labeled buttons again
The problem with the icons was that, allthough they look nice, their
meaning was very hard to guess. Since nobody uses them daily, this
created frustration, especially since tooltips are annoying to access
with a wacom tablet. The labeled buttons are also bigger and thus easier
to hit.

Yes, it is ugly.
Attempt to make it a tiny bit better by adjusting the spacing.