preview: use effective bbox; reduce rescale freq.
Use the effective bounding box for the document: if the user has the
frame turned on, it should be respected.

Reduce the frequency of rescale-and-recenters a bit by only doing it
when the bounds of the document have grown. Prevents some UI juddering.

Catch new widget size allocations and rescale then.