flips: ignore mods; layer move: exit if no mods
Mode flip actions are intended to be bound to keypress shortcuts, where
modifiers are arbitrary and shouldn't imply mode-hold or behavioural
changes. Therefore, instruct modes constructed in flip activations to
ignore the initial set of modifier keys.

Make LayerMoveMode exit after dragging if no modifiers were held when it
was entered, or if launched in response to a mode flip action. This
makes it nicer to bind to unmodified pointer buttons.

https://gna.org/bugs/?20307

This leaves open the question of whether drag-like mode buttons on the
toolbar should persist or exit after their drag, with specific reference
to the expectations of tablet PC users: I'm undecided, but at least we
can now address it as a separate issue.

https://gna.org/bugs/?16869