keyboard events: do not propagate
Fix regression: keyboard events no longer propagate by default. This was
causing some widgets to activate confusingly, giving the appearance of
hangs and nonresponsiveness to spacebar drags.

https://gna.org/bugs/?20326

Derived class handler code is still responsible for declaring when an
event has been logically consumed, but the base class's fallthrough code
has been modified so that the default is to not propagate the event when
falling through to try other mixed-in behaviours.