layer win: nicer toggling of visible & lock
Don't immediately scroll the layers list to the currently highlighted
row if the user just scrolled it to a far-away layer to toggle the
visibility or locked flags on that layer. This allows users to toggle
off a set of far-away layers more conveniently.

https://gna.org/bugs/?20330

However the current layer should remain visible much of the time so
users don't get lost. Experimentally, automatically scroll the list back
to the highlighted layer after each stroke on the canvas, and also after
changing the current layer's opacity or mode inside the dialog. The idea
here is that if you change a layer, you get to see which layer just
changed in the layers list.