drag modes: exit the mode if grabs fail
Another attempt at fixing https://gna.org/bugs/?20068

Falling back to the mode's passive state is proving confusing and
potentially annoying. These conditions should be rare enough for popping
the mode stack to be a valid approach if grabs fail.

Code tidyups: less silly names for the start/stop methods. Ungrabbing
won't work because the conflicting grabs are evidently external to the
app, so remove that attempted workaround. Exclude double and triple
clicks from starting grabs (shouldn't be an issue, but we'll see).  The
keypress handler is now are now fully responsible for starting its drag,
so never start one in a motion handler and clean up test. Eat all motion
events in a grab.