brush list: revert double-click action
After testing it properly with a stylus: it didn't quite have the "feel"
I wanted. Felt right with the mouse, but with the stylus I found I was
doing accidental drags after double-clicks, and accidental double-clicks
during selection. Perhaps tablets "bounce" too much to make double-click
activation a good idea when other stuff is going on.

I think it should be possible to edit properties from the brush list,
but this doesn't look like the right way to do it.